services/billing/requirements.txt
# FastAPI and ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Stripe
stripe>=11.0.0

# Database
supabase>=2.10.0

# Redis
redis>=5.2.0

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# HTTP Client
httpx>=0.28.0

# Utils
python-dotenv>=1.0.1
python-dateutil>=2.9.0

# Logging
structlog>=24.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
